+++ TECH_NOTES.md
@@ -1,17 +1,14 @@
 # Local setup:
 ___
-This repository uses poetry to handle the dependencies.
+This repository uses `poetry` to handle the dependencies.
 The following steps will help you set up the project locally.
 1. Clone the repository.
-2. Create a virtual environment.
-   - `python -m venv .venv`
-3. Use poetry to install the dependencies.
+2. Use poetry to install the dependencies.
    - `poetry install`
-4. Activate the virtual environment.
+3. Activate the virtual environment.
    - `poetry shell`
-5. Command to remove the .venv folder.
-   - MAC: `rm -rf .venv`
-   - Windows: `rmdir /s .venv`
+4. Command to remove the .venv folder.
+   - `poetry env remove`
 
 
 # Python setup using pyenv:
