+++ modo_kit_central/mkc/gui.py
@@ -1,3 +1,6 @@
+"""The main window for Modo Kit Central."""
+from typing import Union, Type
+
 try:
     from PySide6.QtCore import Qt
     from PySide6.QtGui import QCloseEvent, QPixmap
@@ -9,8 +12,9 @@ except ImportError:
     from PySide2.QtWidgets import QMainWindow, QTabWidget, QWidget, QVBoxLayout, QTabBar, QLabel
 
 # Kit imports
-from .prefs import Text, KEYS, DATA, Paths
-from .widgets import Banner, KitsTab, InfoTab
+from .prefs import Text, KEYS, DATA, Paths, TabRequest
+from .widgets.tabs import tab_map, TAB, KitsTab
+from .widgets.core import Banner
 
 
 class KitCentralWindow(QMainWindow):
@@ -23,9 +27,9 @@ class KitCentralWindow(QMainWindow):
         self._build_window()
         self._build_ui()
         self._build_tabs()
+        self._connect_ui()
         # Display the UI
         self.show()
-        # Trigger kit population thread.
 
     def _build_window(self) -> None:
         """Sets up the main window properties."""
@@ -55,34 +59,22 @@ class KitCentralWindow(QMainWindow):
         self.tabs = QTabWidget()
         self.tab_bar = self.tabs.tabBar()
         self.tabs.setTabsClosable(True)
-        self.tabs.tabCloseRequested.connect(self.tab_close)
         self.tabs.setObjectName("Tabs")
         self.base_layout.addWidget(self.tabs)
 
         # Add the core tab
-        self.tab_kits = KitsTab()
-        self.tabs.addTab(self.tab_kits, KEYS.KITS)
-
+        self.tab_kits = self.load_tab(TabRequest(KEYS.KITS, closeable=False))
         # Add the help tab
-        self.tab_help = InfoTab()
-        # Add the help tab but to the right side in self.tabs
-        self.tabs.addTab(self.tab_help, KEYS.INFO)
+        self.tab_help = self.load_tab(TabRequest(KEYS.INFO, closeable=False))
 
-        # Clear all close buttons from the initial tabs.
-        for tab_index in range(self.tabs.count()):
-            self.tab_bar.setTabButton(tab_index, QTabBar.RightSide, None)
-            self.tab_bar.setTabButton(tab_index, QTabBar.LeftSide, None)
+    def _connect_ui(self) -> None:
+        """Connects the UI elements to their respective slots."""
+        # Connect the tab bar to the close event.
+        self.tab_bar.tabCloseRequested.connect(self._tab_close)
+        # Connect the author request to the load tab function.
+        self.tab_kits.author_request.connect(self.load_tab)
 
-    def closeEvent(self, event: QCloseEvent) -> None:
-        """PySide method: Handle closing the UI
-
-        Args:
-            event: The close event from the Window.
-        """
-        self.close()
-        event.accept()
-
-    def tab_close(self, index: int) -> None:
+    def _tab_close(self, index: int) -> None:
         """Handle closing extra tabs.
 
         Args:
@@ -99,3 +91,52 @@ class KitCentralWindow(QMainWindow):
             self.tabs.removeTab(index)
             # Destroy widget as it's no longer needed.
             tab_widget.deleteLater()
+
+    def load_tab(self, tab_request: TabRequest) -> Union[TAB, None]:
+        """Open a new tab in the Kit Central main layout.
+
+        Args:
+            tab_request: The request to open a new tab.
+
+        Returns:
+            QWidget: The new tab widget if it was created, otherwise None.
+        """
+        # Get the intended tab name and check to see if it exists.
+        tab_name = tab_request.name if tab_request.name else tab_request.type
+        tab = self.tabs.findChild(QWidget, tab_name)
+        # If the tab exists, get the index of the tab, else set it to -1.
+        tab_index = self.tabs.indexOf(tab) if tab else -1
+
+        if tab_index != -1:
+            # The tab already exists, so just show it.
+            self.tabs.setCurrentIndex(tab_index)
+            return self.tabs.widget(tab_index)
+
+        # Looks like we need to create a new tab. Let's grab it from the tab map.
+        tab_class: Type[TAB] = tab_map.get(tab_request.type)
+        if not tab_class:
+            return None
+
+        # Initialize the tab with the request data if it exists.
+        new_tab = tab_class(**tab_request.kwargs) if tab_request.kwargs else tab_class()
+        # Add the new tab to the tab bar and get the index.
+        new_tab_index = self.tabs.addTab(new_tab, tab_name)
+
+        if tab_request.show:
+            # If the tab is requested to be shown, set it to the current index.
+            self.tabs.setCurrentIndex(new_tab_index)
+        if not tab_request.closeable:
+            # If the tab is not closeable, set the close button to None.
+            self.tab_bar.setTabButton(new_tab_index, QTabBar.RightSide, None)
+            self.tab_bar.setTabButton(new_tab_index, QTabBar.LeftSide, None)
+
+        return new_tab
+
+    def closeEvent(self, event: QCloseEvent) -> None:
+        """PySide method: Handle closing the UI
+
+        Args:
+            event: The close event from the Window.
+        """
+        self.close()
+        event.accept()
