+++ pyproject.toml
@@ -12,7 +12,7 @@ package-mode = true
 [tool.poetry.dependencies]
 python = ">=3.9,<3.10"
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
 # Add your development dependencies here
 PySide6 = "6.3.1"
 toml = "0.10.2"
